IOS-271 Split into files
