+++ IdentityAccess/IdentityAccessImplementations/persistence/DatabaseTests.swift
@@ -10,6 +10,9 @@ class DatabaseTests: XCTestCase {
     let fm = MockFileManager()
     let sqlite = MockCSQLite3()
     var db: Database!
+    var stmt: Statement!
+    var conn: Connection!
+    var rs: ResultSet!
     let bundleId = "my_random_bundle_id"
 
     enum Error: String, LocalizedError, Hashable {
@@ -64,9 +67,7 @@ class DatabaseTests: XCTestCase {
     }
 
     func test_whenConnecting_thenOpensSqlite() throws {
-        prepareDatabaseConnection()
-        try db.create()
-        _ = try db.connection()
+        try givenConnection()
         XCTAssertEqual(sqlite.openedFilename, db.url.path)
     }
 
@@ -104,19 +105,15 @@ class DatabaseTests: XCTestCase {
     }
 
     func test_connectionClosesSQLiteDatabase() throws {
-        prepareDatabaseConnection()
+        try givenConnection()
         sqlite.close_result = CSQLite3.SQLITE_OK
-        try db.create()
-        let conn = try db.connection()
         try db.close(conn)
         XCTAssertTrue(sqlite.close_pointer == sqlite.open_pointer_result)
     }
 
     func test_whenClosingNotPossible_throwsError() throws {
-        prepareDatabaseConnection()
+        try givenConnection()
         sqlite.close_result = CSQLite3.SQLITE_BUSY
-        try db.create()
-        let conn = try db.connection()
         assertThrows(try db.close(conn), Database.Error.databaseBusy)
         sqlite.close_result = CSQLite3.SQLITE_OK
     }
@@ -132,13 +129,11 @@ class DatabaseTests: XCTestCase {
     }
 
     func test_prepareStatement_passesCorrectArguments() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
+        try givenConnection()
+
         sqlite.prepare_result = CSQLite3.SQLITE_OK
         sqlite.prepare_out_ppStmt = opaquePointer()
         sqlite.prepare_out_pzTail = nil
-
-        let conn = try db.connection()
         _ = try conn.prepare(statement: "some")
 
         XCTAssertEqual(sqlite.prepare_in_db, sqlite.open_pointer_result)
@@ -149,33 +144,23 @@ class DatabaseTests: XCTestCase {
     }
 
     func test_preapreStatement_whenFailed_thenThrowsError() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
+        try givenConnection()
         sqlite.prepare_out_ppStmt = opaquePointer()
         sqlite.prepare_out_pzTail = nil
-
         sqlite.prepare_result = CSQLite3.SQLITE_ERROR
-
-        let conn = try db.connection()
         assertThrows(try conn.prepare(statement: "some"), Database.Error.invalidSQLStatement)
     }
 
     func test_prepareStatement_whenReceivesNilStatement_thenThrowsError() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
+        try givenConnection()
         sqlite.prepare_result = CSQLite3.SQLITE_OK
         sqlite.prepare_out_ppStmt = nil
         sqlite.prepare_out_pzTail = nil
-
-        let conn = try db.connection()
         assertThrows(try conn.prepare(statement: "some"), Database.Error.invalidSQLStatement)
     }
 
     func test_whenConnectionIsClosed_thenPreparedStatementIsFinalizedAutomatically() throws {
-        try db.create()
-
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
+        try givenConnection()
 
         sqlite.prepare_result = CSQLite3.SQLITE_OK
         sqlite.prepare_out_ppStmt = opaquePointer()
@@ -189,10 +174,7 @@ class DatabaseTests: XCTestCase {
     }
 
     func test_whenMultipleStatementsCreated_finalizesAll() throws {
-        try db.create()
-
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
+        try givenConnection()
 
         let stmt1 = opaquePointer()
         sqlite.prepare_result = CSQLite3.SQLITE_OK
@@ -210,7 +192,6 @@ class DatabaseTests: XCTestCase {
         try db.close(conn)
 
         XCTAssertEqual(sqlite.finalize_in_pStmt_list, [stmt1, stmt2])
-
     }
 
     func test_whenStatementFinalizedAndExecutes_thenThrows() throws {
@@ -220,107 +201,60 @@ class DatabaseTests: XCTestCase {
     }
 
     func test_whenConnectionWasClosedAndThenOpened_thenThrows() throws {
-        try db.create()
-
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-
-        try db.close(conn)
-
+        try givenClosedConnection()
         assertThrows(try conn.open(url: db.url), Database.Error.connectionIsAlreadyClosed)
     }
 
     func test_whenConnectionWasClosedAndThenClosedAgain_thenThrows() throws {
-        try db.create()
-
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-
-        try db.close(conn)
-
+        try givenClosedConnection()
         assertThrows(try db.close(conn), Database.Error.connectionIsAlreadyClosed)
     }
 
+
     func test_whenConnectionClosed_thenPrepareThrows() throws {
-        try db.create()
-
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-
-        try db.close(conn)
+        try givenClosedConnection()
         assertThrows(try conn.prepare(statement: "some"), Database.Error.connectionIsAlreadyClosed)
     }
 
-    func test_whenDatabaseDestroyed_thenConnectionsAreClosed() throws {
-
-    }
-
     func test_whenDatabaseDeinit_thenConnectionsAreClosed() throws {
-        try db.create()
-
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-
+        try givenConnection()
         try db.destroy()
-
         assertThrows(try conn.prepare(statement: "some"), Database.Error.connectionIsAlreadyClosed)
     }
 
     func test_canCreateAfterDestroy() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        _ = try db.connection()
+        try givenConnection()
         try db.destroy()
         try db.create()
         try db.destroy()
     }
 
     func test_execute_whenStepReturnsDone_thenOk() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
+        try givenPreparedStatement()
         try stmt.execute()
         XCTAssertNotNil(sqlite.step_in_pStmt)
     }
 
     func test_whenExecutedAndExecutesAgain_thenThrows() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
+        try givenPreparedStatement()
         try stmt.execute()
         assertThrows(try stmt.execute(), Database.Error.statementWasAlreadyExecuted)
     }
 
     func test_whenExecuteError_thenThrows() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
+        try givenPreparedStatement()
         sqlite.step_results = [CSQLite3.SQLITE_ERROR]
         assertThrows(try stmt.execute(), Database.Error.runtimeError)
     }
 
     func test_whenExecuteMisuse_thenTrows() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
+        try givenPreparedStatement()
         sqlite.step_results = [CSQLite3.SQLITE_MISUSE]
         assertThrows(try stmt.execute(), Database.Error.invalidStatementState)
     }
 
     func test_whenStatementIsNotCommitAndOccursInsideExplicitTransaction_thenThrows() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
+        try givenPreparedStatement()
         sqlite.step_results = [CSQLite3.SQLITE_BUSY]
         // 0 means False - autocommit is disabled - inside BEGIN...COMMIT
         sqlite.get_autocommit_result = 0
@@ -328,22 +262,14 @@ class DatabaseTests: XCTestCase {
     }
 
     func test_whenStatementIsCommitAndBusy_thenRetries() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "COMMIT;")
+        try givenPreparedStatement("COMMENT;")
         sqlite.step_results = [CSQLite3.SQLITE_BUSY, CSQLite3.SQLITE_BUSY, CSQLite3.SQLITE_DONE]
         try stmt.execute()
         XCTAssertEqual(sqlite.step_result_index, sqlite.step_results.count)
     }
 
     func test_whenStatementNotCommitAndBusyAndOutsideExplicitTransaction_thenRetries() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
+        try givenPreparedStatement()
         sqlite.step_results = [CSQLite3.SQLITE_BUSY, CSQLite3.SQLITE_BUSY, CSQLite3.SQLITE_DONE]
         // autocommit enabled - means not in BEGIN...COMMIT
         sqlite.get_autocommit_result = 1
@@ -351,50 +277,23 @@ class DatabaseTests: XCTestCase {
         XCTAssertEqual(sqlite.step_result_index, sqlite.step_results.count)
     }
 
-    func test_whenStatementProducesResult_thenReturnsResultSet() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
-        sqlite.step_results = [CSQLite3.SQLITE_ROW]
-        let rs = try stmt.execute()
-        XCTAssertNotNil(rs)
-    }
-
     func test_whenStatementHasNoColumns_thenReturns() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
-        sqlite.step_results = [CSQLite3.SQLITE_ROW]
-        guard let rs = try stmt.execute() else { throw Error.resultSetMissing }
+        try givenResultSet()
         sqlite.column_count_result = 0
         XCTAssertTrue(rs.isColumnsEmpty)
         XCTAssertNotNil(sqlite.column_count_in_pStmt)
     }
 
     func test_resultSetColumnCount() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
-        sqlite.step_results = [CSQLite3.SQLITE_ROW]
-        guard let rs = try stmt.execute() else { throw Error.resultSetMissing }
+        try givenResultSet()
         sqlite.column_count_result = 15
         XCTAssertEqual(rs.columnCount, 15)
     }
 
+
     func test_resultSet_columnValues() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
-        sqlite.step_results = [CSQLite3.SQLITE_ROW]
-        guard let rs = try stmt.execute() else { throw Error.resultSetMissing }
+        try givenResultSet()
+
         sqlite.column_count_result = 3
         sqlite.column_text_result = "some"
         sqlite.column_bytes_result = Int32("some".cString(using: .utf8)!.count)
@@ -408,24 +307,14 @@ class DatabaseTests: XCTestCase {
     }
 
     func test_whenReturnsRow_thenResetsQuery() throws {
-        // query reset so that ResultSet.advanceToNextRow() can be used
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
+        try givenPreparedStatement()
         sqlite.step_results = [CSQLite3.SQLITE_ROW]
         try stmt.execute()
         XCTAssertNotNil(sqlite.reset_in_pStmt)
     }
 
     func test_whenReturns2Rows_thenCanAdvance2Times() throws {
-        // query reset so that ResultSet.advanceToNextRow() can be used
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
+        try givenPreparedStatement()
         sqlite.step_results = [CSQLite3.SQLITE_ROW, // called by execute()
             CSQLite3.SQLITE_ROW, // advanceToNextRow()
             CSQLite3.SQLITE_ROW,
@@ -437,11 +326,7 @@ class DatabaseTests: XCTestCase {
     }
 
     func test_whenOutsideOfExplicitTransactionAndBusyDuringAdvancing_thenRetries() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
+        try givenPreparedStatement()
         sqlite.get_autocommit_result = 1
         sqlite.step_results = [CSQLite3.SQLITE_ROW, // called by execute()
             CSQLite3.SQLITE_ROW, // advanceToNextRow()
@@ -455,12 +340,9 @@ class DatabaseTests: XCTestCase {
         XCTAssertFalse(try rs.advanceToNextRow())
     }
 
-    func test_statementSetText() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
+    func test_statement_setParameters() throws {
+        try givenPreparedStatement()
+
         sqlite.bind_text_result = CSQLite3.SQLITE_OK
         try stmt.set("text", at: 1)
         XCTAssertEqual(sqlite.bind_text_in_zValue, "text")
@@ -502,14 +384,9 @@ class DatabaseTests: XCTestCase {
         XCTAssertEqual(sqlite.bind_parameter_index_in_zName, "key")
     }
 
+
     func test_whenStatementExecuted_thenBindingThrows() throws {
-        try db.create()
-        sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
-        sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
-        sqlite.step_results = [CSQLite3.SQLITE_DONE]
-        _ = try stmt.execute()
+        try givenSuccessfullyExecutedStatement()
         assertThrows(try stmt.set(0, at: 1), Database.Error.attemptToBindExecutedStatement)
         assertThrows(try stmt.set(1.1, at: 1), Database.Error.attemptToBindExecutedStatement)
         assertThrows(try stmt.set("text", at: 1), Database.Error.attemptToBindExecutedStatement)
@@ -520,13 +397,17 @@ class DatabaseTests: XCTestCase {
         assertThrows(try stmt.setNil(forKey: "key"), Database.Error.attemptToBindExecutedStatement)
     }
 
-    func test_whenStatementFinalized_thenBindingThrows() throws {
+    fileprivate func givenNotExecutedFinalizedStatement() throws {
         try db.create()
         sqlite.open_pointer_result = opaquePointer()
-        let conn = try db.connection()
+        conn = try db.connection()
         sqlite.prepare_out_ppStmt = opaquePointer()
-        let stmt = try conn.prepare(statement: "some")
+        stmt = try conn.prepare(statement: "some")
         stmt.finalize()
+    }
+
+    func test_whenStatementFinalized_thenBindingThrows() throws {
+        try givenNotExecutedFinalizedStatement()
         assertThrows(try stmt.set(0, at: 1), Database.Error.attemptToBindFinalizedStatement)
         assertThrows(try stmt.set(1.1, at: 1), Database.Error.attemptToBindFinalizedStatement)
         assertThrows(try stmt.set("text", at: 1), Database.Error.attemptToBindFinalizedStatement)
@@ -540,9 +421,44 @@ class DatabaseTests: XCTestCase {
 }
 
 extension DatabaseTests {
-
-    private func prepareDatabaseConnection() {
+    fileprivate func givenSuccessfullyExecutedStatement() throws {
+        try db.create()
         sqlite.open_pointer_result = opaquePointer()
+        conn = try db.connection()
+        sqlite.prepare_out_ppStmt = opaquePointer()
+        stmt = try conn.prepare(statement: "some")
+        sqlite.step_results = [CSQLite3.SQLITE_DONE]
+        _ = try stmt.execute()
+    }
+
+    fileprivate func givenPreparedStatement(_ sql: String = "some") throws {
+        try db.create()
+        sqlite.open_pointer_result = opaquePointer()
+        conn = try db.connection()
+        sqlite.prepare_out_ppStmt = opaquePointer()
+        stmt = try conn.prepare(statement: sql)
+    }
+
+    fileprivate func givenClosedConnection() throws {
+        try givenConnection()
+        try db.close(conn)
+    }
+
+    fileprivate func givenConnection() throws {
+        try db.create()
+        sqlite.open_pointer_result = opaquePointer()
+        conn = try db.connection()
+    }
+
+    fileprivate func givenResultSet() throws {
+        try db.create()
+        sqlite.open_pointer_result = opaquePointer()
+        conn = try db.connection()
+        sqlite.prepare_out_ppStmt = opaquePointer()
+        stmt = try conn.prepare(statement: "some")
+        sqlite.step_results = [CSQLite3.SQLITE_ROW]
+        guard let rs = try stmt.execute() else { throw Error.resultSetMissing }
+        self.rs = rs
     }
 
     func opaquePointer() -> OpaquePointer {
@@ -559,34 +475,5 @@ extension DatabaseTests {
         XCTAssertThrowsError(expression, file: file, line: line) {
             XCTAssertEqual($0 as? E, error, file: file, line: line)
         }
-    }
-}
-
-class MockFileManager: FileManager {
-
-    func appSupportURL() throws -> URL {
-        return try url(for: .applicationSupportDirectory,
-                       in: .userDomainMask,
-                       appropriateFor: nil,
-                       create: false)
-    }
-
-    var notExistingURLs = [URL]()
-    var existingURLs = [URL]()
-
-    override func fileExists(atPath path: String) -> Bool {
-        if existingURLs.map({ $0.path }).contains(path) { return true }
-        if notExistingURLs.map({ $0.path }).contains(path) { return false }
-        return super.fileExists(atPath: path)
-    }
-
-    override func removeItem(at URL: URL) throws {
-        if let index = existingURLs.index(of: URL) {
-            existingURLs.remove(at: index)
-            if !super.fileExists(atPath: URL.path) {
-                return
-            }
-        }
-        try super.removeItem(at: URL)
     }
 }
